update go 1.27.1, fixed image type, update import device, make file upload to 10 mb, fix bug en-th, fix time widgets, add searchbar option, update all protocol device

+++ database/init.sql
@@ -194,7 +194,7 @@ CREATE TABLE IF NOT EXISTS device_rule_notification (
     rule_id SERIAL,
     device_id INT NOT NULL,
     condition TEXT NOT NULL, -- e.g., '>', '<', '==', '>=', '<='
-    threshold INT NOT NULL,
+    threshold DOUBLE PRECISION NOT NULL,
     reason TEXT,
     active BOOLEAN NOT NULL DEFAULT TRUE,
     created_at TIMESTAMPTZ NOT NULL DEFAULT now(),
